requirements.txt
pytest==8.4.1
requests==2.32.4
python-dotenv==1.0.1
assertpy==1.1
pytest-html-plus==0.4.4